+++ src/pid_controller.h
@@ -7,14 +7,14 @@
 class PidController{
 public:
     PidController();//the constructor could get the PID params from the eeprom. the setter of kP,kI,kD could store the values in eeprom
-    int16_t update(int32_t value);
+    int16_t update(int16_t value);
 
-    int32_t sum;
-    int32_t last_diff;
-    int32_t target;
-    uint32_t kP;
-    uint32_t kD;
-    uint32_t kI;
+    int16_t sum;
+    int16_t last_diff;
+    int16_t target;
+    uint16_t kP;
+    uint16_t kD;
+    uint16_t kI;
     int16_t maximum = 1000;
-    int32_t sum_max;
+    int16_t sum_max;
 };
